Rename compiler to compile-expression
